DATAGRAPH-1425 - Support built-in strict mode.
