+++ src/test/java/org/springframework/data/neo4j/integration/kotlin/KotlinIT.java
@@ -27,10 +27,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.neo4j.config.AbstractNeo4jConfig;
-import org.springframework.data.neo4j.integration.shared.KotlinClub;
-import org.springframework.data.neo4j.integration.shared.KotlinClubRelationship;
-import org.springframework.data.neo4j.integration.shared.KotlinPerson;
-import org.springframework.data.neo4j.integration.shared.KotlinRepository;
+import org.springframework.data.neo4j.integration.shared.common.KotlinClub;
+import org.springframework.data.neo4j.integration.shared.common.KotlinClubRelationship;
+import org.springframework.data.neo4j.integration.shared.common.KotlinPerson;
+import org.springframework.data.neo4j.integration.shared.common.KotlinRepository;
 import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 import org.springframework.data.neo4j.test.Neo4jExtension.Neo4jConnectionSupport;
 import org.springframework.data.neo4j.test.Neo4jIntegrationTest;
