+++ src/test/kotlin/org/springframework/data/neo4j/integration/shared/common/ImmutableRelationships.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2011-2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.data.neo4j.integration.shared.common
+
+import org.springframework.data.neo4j.core.schema.GeneratedValue
+import org.springframework.data.neo4j.core.schema.Id
+import org.springframework.data.neo4j.core.schema.Node
+import org.springframework.data.neo4j.core.schema.Relationship
+
+/**
+ * @author Gerrit Meier
+ */
+@Node
+data class DeviceEntity(
+        @Id
+        val deviceId: String,
+        val phoneNumber: String,
+        @Relationship(type = "LATEST_LOCATION", direction = Relationship.Direction.OUTGOING)
+        val location: LocationEntity?
+)
+
+@Node
+data class LocationEntity(
+        @Id
+        @GeneratedValue
+        val locationId: Long? = null,
+        val latitude: Double,
+        val longitude: Double,
+        @Relationship(type = "PREVIOUS_LOCATION", direction = Relationship.Direction.OUTGOING)
+        val previousLocation: LocationEntity?
+)
+
+@Node
+data class ImmutableKotlinPerson(@Id val name: String, val wasOnboardedBy: List<ImmutableKotlinPerson>)
