+++ src/test/kotlin/org/springframework/data/neo4j/integration/shared/common/KotlinPerson.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.data.neo4j.integration.shared
+package org.springframework.data.neo4j.integration.shared.common
 
 import org.springframework.data.neo4j.core.schema.GeneratedValue
 import org.springframework.data.neo4j.core.schema.Id
